Split into Impl Bundle and APi Bundle
