+++ eim.api/.settings/org.eclipse.core.resources.prefs
@@ -1,5 +1,4 @@
 eclipse.preferences.version=1
-encoding//root/plugin.xml=UTF-8
+encoding//src/eim/api/package-info.java=UTF-8
 encoding/<project>=UTF-8
 encoding/bnd.bnd=UTF-8
-encoding/eim.bndrun=UTF-8
