+++ eim.api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Version: 0.1.0.${tstamp}
+
+-baseline: *
+-buildpath: \
+  osgi.annotation; version=7.0.0
+
