+++ eim.api/src/eim/api/EclipseService.java
@@ -1,5 +1,8 @@
-package eim;
+package eim.api;
 
+import org.osgi.annotation.versioning.ProviderType;
+
+@ProviderType
 public interface EclipseService {
 	void startProcess(String command, String workingDir, String[] args);
 }
