+++ eim/command.bnd
@@ -1,0 +1,1 @@
+Private-Package: command
