+++ eim/eim.bndrun
@@ -1,4 +1,4 @@
--runee: JavaSE-11
+-runee: JavaSE-17
 -runfw: org.eclipse.osgi;version='[3.16,4)'
 
 -runproperties: \
@@ -8,16 +8,19 @@
     eclipse.application=eim.application
 
 -runrequires: \
-	bnd.identity;id=eim;version='[0.1.0,1.0.0)',\
 	bnd.identity;id='org.apache.felix.gogo.command',\
-	bnd.identity;id='org.apache.felix.gogo.shell'
+	bnd.identity;id='org.apache.felix.gogo.shell',\
+	bnd.identity;id='eim.command',\
+	bnd.identity;id='eim.impl'
 -runproperties: osgi.console=
 -runbundles: \
-	eim;version=snapshot,\
 	org.apache.felix.gogo.command;version='[1.1.2,1.1.3)',\
 	org.apache.felix.gogo.runtime;version='[1.1.4,1.1.5)',\
 	org.apache.felix.gogo.shell;version='[1.1.4,1.1.5)',\
-	org.apache.felix.scr;version='[2.1.24,2.1.25)',\
 	org.eclipse.osgi.services;version='[3.10.200,3.10.201)',\
 	org.osgi.util.function;version='[1.2.0,1.2.1)',\
-	org.osgi.util.promise;version='[1.2.0,1.2.1)'
+	org.osgi.util.promise;version='[1.2.0,1.2.1)',\
+	eim.api;version=snapshot,\
+	eim.command;version=snapshot,\
+	eim.impl;version=snapshot,\
+	org.apache.felix.scr;version='[2.1.24,2.1.25)'
