+++ eim/src/command/EclipseStarterCommand.java
@@ -1,10 +1,16 @@
-package eim;
+package command;
 
+
+import org.apache.felix.service.command.Descriptor;
+import org.osgi.annotation.versioning.ConsumerType;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-@Component(property = { "osgi.command.scope:String=zEMI",
-		"osgi.command.function:String=startProcess" }, service = EclipseStarterCommand.class)
+import eim.api.EclipseService;
+
+@Component(property = { "osgi.command.scope=zEMI",  
+		"osgi.command.function=startProcess" }, service = EclipseStarterCommand.class)
+@ConsumerType
 public class EclipseStarterCommand {
 
 	private EclipseService eclService;
@@ -13,7 +19,8 @@ public class EclipseStarterCommand {
 	public void bindEclipseService(EclipseService eclsvc) {
 		this.eclService = eclsvc;
 	}
-
+	
+	@Descriptor("Start a process with a specific command.")
 	public void startProcess(String command, String workingDir, String[] args) {
 		System.out.println("Executing " + command);
 		eclService.startProcess(command, workingDir, args);
