--- eim/src/eim/package-info.java
@@ -1,1 +1,0 @@
-package eim;
